Provides an overview of all resources available on the Open Data platform.

Provides an overview of all resources available from opendata.nhs.scot, with the option to limit results based on both dataset and resource names. The returned tibble can be used to look-up dataset and resource ids, and is useful for exploring the available data sets.

list_resources(
  dataset_contains = NULL,
  resource_contains = NULL,
  dataset_name = lifecycle::deprecated()
)

Arguments

dataset_contains

A character string containing an expression to be used as search criteria against the dataset name.

resource_contains

A character string containing a regular expression to be matched against available resource names.

dataset_name

Deprecated. Use dataset_contains instead.

Value

A tibble containing details of all available datasets and resources, or those containing the string specified in the dataset_contains and resource_contains arguments.

Examples

list_resources()
#> # A tibble: 1,115 × 7
#>    resource_name         resource_id dataset_name dataset_title dataset_id url  
#>    <chr>                 <chr>       <chr>        <chr>         <chr>      <chr>
#>  1 "Weekly Cases for Al… d08c0fbd-7… viral-respi… Viral Respir… 49dc2d88-… http…
#>  2 "Weekly Cases for CO… 212412ba-c… viral-respi… Viral Respir… 49dc2d88-… http…
#>  3 "Weekly Cases for Al… 170ca9d4-8… viral-respi… Viral Respir… 49dc2d88-… http…
#>  4 "Weekly Cases for CO… 7ca4f902-7… viral-respi… Viral Respir… 49dc2d88-… http…
#>  5 "Weekly Cases for CO… 9c0f7af4-f… viral-respi… Viral Respir… 49dc2d88-… http…
#>  6 "Weekly Admissions i… 35115613-9… viral-respi… Viral Respir… 49dc2d88-… http…
#>  7 "Weekly Admissions b… f2ae2048-c… viral-respi… Viral Respir… 49dc2d88-… http…
#>  8 "Weekly Admissions b… 29a6b2a7-5… viral-respi… Viral Respir… 49dc2d88-… http…
#>  9 " Weekly Admissions … 4715ed6c-0… viral-respi… Viral Respir… 49dc2d88-… http…
#> 10 " Weekly tests and t… 25ad967b-e… viral-respi… Viral Respir… 49dc2d88-… http…
#> # ℹ 1,105 more rows
#> # ℹ 1 more variable: last_modified <dttm>
list_resources(dataset_contains = "standard-populations")
#> # A tibble: 4 × 7
#>   resource_name          resource_id dataset_name dataset_title dataset_id url  
#>   <chr>                  <chr>       <chr>        <chr>         <chr>      <chr>
#> 1 European Standard Pop… edee9731-d… standard-po… Standard Pop… 4dd86111-… http…
#> 2 European Standard Pop… 29ce4cda-a… standard-po… Standard Pop… 4dd86111-… http…
#> 3 World Standard Popula… e27533df-a… standard-po… Standard Pop… 4dd86111-… http…
#> 4 World Standard Popula… 2f493d21-f… standard-po… Standard Pop… 4dd86111-… http…
#> # ℹ 1 more variable: last_modified <dttm>
list_resources(
  dataset_contains = "standard-populations", resource_contains = "European"
)
#> # A tibble: 2 × 7
#>   resource_name          resource_id dataset_name dataset_title dataset_id url  
#>   <chr>                  <chr>       <chr>        <chr>         <chr>      <chr>
#> 1 European Standard Pop… edee9731-d… standard-po… Standard Pop… 4dd86111-… http…
#> 2 European Standard Pop… 29ce4cda-a… standard-po… Standard Pop… 4dd86111-… http…
#> # ℹ 1 more variable: last_modified <dttm>