Get Open Data resources from a dataset

Get Open Data resources from a dataset

Usage

get_dataset(
  dataset_name,
  max_resources = NULL,
  rows = NULL,
  row_filters = NULL,
  col_select = NULL,
  include_context = FALSE
)

Arguments

dataset_name

name of the dataset as found on NHS Open Data platform

max_resources

(optional) the maximum number of resources to return, use for testing code, it will return the n latest resources

rows

(optional) specify the max number of rows to return for each resource.

row_filters

(optional) a named list or vector that specifies values of columns/fields to keep. e.g. list(Date = 20220216, Sex = "Female").

col_select

(optional) a character vector containing the names of desired columns/fields. e.g. c("Date", "Sex").

include_context

(optional) If TRUE additional information about the resource will be added as columns to the data, including the resource ID, the resource name, the creation date and the last modified/updated date.

Value

a tibble with the data

See also

Examples

get_dataset("gp-practice-populations",
  max_resources = 2, rows = 10
)
#> # A tibble: 20 × 24
#>        Date PracticeCode HB        HSCP   Sex   SexQF AllAges AllAgesQF Ages0to4
#>       <int>        <int> <chr>     <chr>  <chr> <chr>   <int> <chr>        <int>
#>  1 20250101        10002 S08000030 S3700… Male  ""       4118 ""             168
#>  2 20250101        10002 S08000030 S3700… Fema… ""       4248 ""             170
#>  3 20250101        10002 S08000030 S3700… All   "d"      8366 ""             338
#>  4 20250101        10017 S08000030 S3700… Male  ""       3778 ""             133
#>  5 20250101        10017 S08000030 S3700… Fema… ""       3778 ""             123
#>  6 20250101        10017 S08000030 S3700… All   "d"      7556 ""             256
#>  7 20250101        10036 S08000030 S3700… Male  ""       2446 ""              90
#>  8 20250101        10036 S08000030 S3700… Fema… ""       2406 ""              54
#>  9 20250101        10036 S08000030 S3700… All   "d"      4852 ""             144
#> 10 20250101        10106 S08000030 S3700… Male  ""       3092 ""              92
#> 11 20241001        10002 S08000030 S3700… Male  ""       4109 ""             173
#> 12 20241001        10002 S08000030 S3700… Fema… ""       4221 ""             172
#> 13 20241001        10002 S08000030 S3700… All   "d"      8330 ""             345
#> 14 20241001        10017 S08000030 S3700… Male  ""       3742 ""             136
#> 15 20241001        10017 S08000030 S3700… Fema… ""       3721 ""             119
#> 16 20241001        10017 S08000030 S3700… All   "d"      7463 ""             255
#> 17 20241001        10036 S08000030 S3700… Male  ""       2452 ""              96
#> 18 20241001        10036 S08000030 S3700… Fema… ""       2412 ""              56
#> 19 20241001        10036 S08000030 S3700… All   "d"      4864 ""             152
#> 20 20241001        10106 S08000030 S3700… Male  ""       3106 ""              96
#> # ℹ 15 more variables: Ages0To4QF <chr>, Ages5to14 <int>, Ages5To14QF <chr>,
#> #   Ages15to24 <int>, Ages15To24QF <chr>, Ages25to44 <int>, Ages25To44QF <chr>,
#> #   Ages45to64 <int>, Ages45To64QF <chr>, Ages65to74 <int>, Ages65To74QF <chr>,
#> #   Ages75to84 <int>, Ages75To84QF <chr>, Ages85plus <int>, Ages85PlusQF <chr>