Get Open Data resources from a dataset

Get Open Data resources from a dataset

Usage

get_dataset(dataset_name, max_resources = NULL, rows = NULL)

Arguments

dataset_name

name of the dataset as found on NHS Open Data platform

max_resources

(optional) the maximum number of resources to return, use for testing code, it will return the n latest resources

rows

(optional) specify the max number of rows to return for each resource.

Value

a tibble with the data

See also

Examples

get_dataset("gp-practice-populations",
  max_resources = 2, rows = 10
)
#> # A tibble: 20 × 24
#>        Date PracticeCode HB        HSCP   Sex   SexQF AllAges AllAgesQF Ages0to4
#>       <int>        <int> <chr>     <chr>  <chr> <chr>   <int> <chr>        <int>
#>  1 20240101        10002 S08000030 S3700… Male  ""       4065 ""             190
#>  2 20240101        10002 S08000030 S3700… Fema… ""       4209 ""             171
#>  3 20240101        10002 S08000030 S3700… All   "d"      8274 ""             361
#>  4 20240101        10017 S08000030 S3700… Male  ""       3596 ""             126
#>  5 20240101        10017 S08000030 S3700… Fema… ""       3568 ""             113
#>  6 20240101        10017 S08000030 S3700… All   "d"      7164 ""             239
#>  7 20240101        10036 S08000030 S3700… Male  ""       2429 ""              99
#>  8 20240101        10036 S08000030 S3700… Fema… ""       2402 ""              53
#>  9 20240101        10036 S08000030 S3700… All   "d"      4831 ""             152
#> 10 20240101        10106 S08000030 S3700… Male  ""       3101 ""              89
#> 11 20231001        10002 S08000030 S3700… Male  ""       4085 ""             186
#> 12 20231001        10002 S08000030 S3700… Fema… ""       4216 ""             170
#> 13 20231001        10002 S08000030 S3700… All   "d"      8301 ""             356
#> 14 20231001        10017 S08000030 S3700… Male  ""       3540 ""             128
#> 15 20231001        10017 S08000030 S3700… Fema… ""       3506 ""             113
#> 16 20231001        10017 S08000030 S3700… All   "d"      7046 ""             241
#> 17 20231001        10036 S08000030 S3700… Male  ""       2445 ""              94
#> 18 20231001        10036 S08000030 S3700… Fema… ""       2429 ""              58
#> 19 20231001        10036 S08000030 S3700… All   "d"      4874 ""             152
#> 20 20231001        10106 S08000030 S3700… Male  ""       3109 ""              96
#> # ℹ 15 more variables: Ages0To4QF <chr>, Ages5to14 <int>, Ages5To14QF <chr>,
#> #   Ages15to24 <int>, Ages15To24QF <chr>, Ages25to44 <int>, Ages25To44QF <chr>,
#> #   Ages45to64 <int>, Ages45To64QF <chr>, Ages65to74 <int>, Ages65To74QF <chr>,
#> #   Ages75to84 <int>, Ages75To84QF <chr>, Ages85plus <int>, Ages85PlusQF <chr>